requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.22,<1.0
python-dotenv>=1.0.0,<2.0
requests>=2.31,<3.0
PyJWT>=2.8,<3.0
email-validator>=2.1,<3.0
passlib[bcrypt]>=1.7,<2.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
mysqlclient>=2.2,<3.0

# NLP / DL
torch>=2.2,<3.0
transformers>=4.41,<5.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.2,<3.0
yake>=0.4.8
huggingface_hub>=0.23,<1.0
beautifulsoup4>=4.12,<5.0
lxml>=4.9,<5.0

# Google Gemini AI
google-generativeai>=0.8.0,<1.0

# Document parsing & PDF
pymupdf>=1.23,<2.0
pdfplumber>=0.11,<1.0
python-docx>=1.1,<2.0
reportlab>=4.0,<5.0
pdf2image>=1.17,<2.0
pytesseract>=0.3.10,<0.4
pillow>=10.0,<11.0
camelot-py>=0.11,<1.0
tabula-py>=2.9,<3.0
jpype1>=1.5,<2.0
pandas>=2.2,<3.0
opencv-python>=4.8,<5.0

# Tip: If PyTorch doesn't install automatically on Windows, install a compatible build manually (CPU exa
#   pip install torch --index-url https://download.pytorch.org/whl/cpu

